more

+++ dds/sqs.py
@@ -1,37 +1,47 @@
 import datetime
+import glob
 import os
 import time
 import subprocess as sp
 import json
 import boto3
 from dds.logs import lg_sqs as lg
-from mat.ddh_shared import dds_get_json_mac_dns
+from dds.sns import sns_serve
+from mat.ddh_shared import dds_get_json_mac_dns, \
+    get_dds_folder_path_sqs, \
+    ddh_get_commit, \
+    dds_get_commit, \
+    dds_get_json_vessel_name
 
 
-# todo > do SQS as SNS based on files after testing
+# todo > maybe add a logger OK SQS message
 
-WS_DDH_BOOTED = 'DDH_BOOTED'
-WS_DDH_ERROR_BLE_HARDWARE = 'DDH_ERROR_BLE_HARDWARE'
-WS_LOGGER_DOWNLOAD = 'LOGGER_DOWNLOAD'
-WS_LOGGER_ERROR_OXYGEN = 'LOGGER_ERROR_OXYGEN'
-WS_LOGGER_ERRORS_MAX = 'LOGGER_ERRORS_MAXED_RETRIES'
-WS_DDH_PRESENT_SET_XP_TIME = 5
-WS_QUEUE_IN_NAME = 'ddw_in'
-g_last_sqs_notify_error_ble_hw = time.perf_counter()
 
+# -----------------
+# SQS credentials
+# -----------------
+from mat.utils import linux_is_rpi3, linux_is_rpi4
 
 session = boto3.Session(
     aws_access_key_id=os.getenv('DDH_AWS_KEY_ID'),
     aws_secret_access_key=os.getenv('DDH_AWS_SECRET')
 )
 sqs = session.resource('sqs', region_name='us-east-2')
+SQS_UP_QUEUE_NAME = os.getenv('DDH_SQS_UP_QUEUE_NAME')
+SQS_PROJECT_NAME = os.getenv('DDH_AWS_PROJECT_NAME')
 
 
-def _sqs_get_queue(name):
-    return sqs.get_queue_by_name(QueueName=name)
+# ------------------
+# SQS message types
+# ------------------
+SQS_BOOT = 'DDH_BOOT'
+SQS_ERROR_BLE_HW = 'DDH_ERROR_BLE_HARDWARE'
+SQS_LOGGER_DL = 'LOGGER_DOWNLOAD'
+SQS_LOGGER_ERROR_OXYGEN = 'LOGGER_ERROR_OXYGEN'
+SQS_LOGGER_MAX_ERRORS = 'LOGGER_ERRORS_MAXED_RETRIES'
 
 
-def _sqs_enqueue_msg(q, m_body, m_attr=None):
+def _msg_queue(q, m_body, m_attr=None):
     if not m_attr:
         m_attr = {}
 
@@ -41,71 +51,80 @@ def _sqs_enqueue_msg(q, m_body, m_attr=None):
     )
 
 
-def _sqs_dequeue_msg(q, n, t):
+def _msg_dequeue(q, n=10, t=5):
 
+    # n == max but if t is small, may get only 1
     msgs = q.receive_messages(
         MessageAttributeNames=['All'],
         MaxNumberOfMessages=n,
         WaitTimeSeconds=t
     )
-    for m in msgs:
-        lg.a('rx msg {}: {}'.format(m.message_id, m.body))
+    s = '[ SQS ] asked de-queuing {} msgs, got {}'
+    print(s.format(n, len(msgs)))
     return msgs
 
 
-def _sqs_del_msg_from_queue(m):
-    """ clients must delete message after they receive them """
+def sqs_ws():
+    q = sqs.get_queue_by_name(QueueName=SQS_UP_QUEUE_NAME)
+    msgs = _msg_dequeue(q)
 
-    m.delete()
-    lg.a('deleted message {}'.format(m.message_id))
+    # forward as SNS
+    for m in msgs:
+        d = json.loads(m.body)
+        i = m.message_id
+        print('[ SQS_WS ] forwarding msg {} to SNS'.format(i))
+        if sns_serve(d) == 0:
+            # must be done by client after de-queuing
+            m.delete()
+        else:
+            print('[ SQS_WS ] could not SNS msg {}'.format(i))
 
 
-def _sqs_build_json_msg(reason, lat, lon) -> str:
-
-    com_ddh = 'v.1234'
-    com_dds = 'v.4567'
-    p = 'osu'
-    v = 'boat_test'
+def _msg_build(description, lat, lon):
     t = datetime.datetime.now().strftime('%Y-%m-%d %H:%M:%S')
     u = datetime.datetime.utcnow().strftime('%Y-%m-%d %H:%M:%S')
-    ddh_box_sn = 'sn1234567'
     rv_up = sp.run('uptime', shell=True, stdout=sp.PIPE, stderr=sp.PIPE)
+    plat = 'dev'
+    if linux_is_rpi3():
+        plat = 'rpi3'
+    elif linux_is_rpi4():
+        plat = 'rpi4'
     d = {
-        'reason': reason,
-        'project': p,
-        'vessel': v,
-        'ddh_commit': com_ddh,
-        'dds_commit': com_dds,
+        'reason': description,
+        'project': SQS_PROJECT_NAME,
+        'vessel': dds_get_json_vessel_name(),
+        'ddh_commit': ddh_get_commit(),
+        'dds_commit': dds_get_commit(),
         'utc_time': str(u),
         'local_time': str(t),
-        'box_sn': ddh_box_sn,
+        'box_sn': os.getenv('DDH_BOX_SERIAL_NUMBER'),
         'hw_uptime': rv_up.stdout.decode(),
-        'gps_position': '{},{}'.format(lat, lon)
+        'gps_position': '{},{}'.format(lat, lon),
+        'platform': plat
     }
-    return json.dumps(d)
+
+    # save as file
+    fol = str(get_dds_folder_path_sqs())
+    now = int(time.time())
+    path = '{}/{}.sqs'.format(fol, now)
+    with open(path, 'w') as f:
+        json.dump(d, f)
+
+    # log it happened
+    s = 'generated {} -> {} at {}, {}'
+    lg.a(s.format(path, description, lat, lon))
 
 
 def sqs_notify_ddh_booted(lat, lon):
     try:
-        q = _sqs_get_queue(WS_QUEUE_IN_NAME)
-        m = _sqs_build_json_msg(WS_DDH_BOOTED, lat, lon)
-        lg.a('<- DDS booted', lat, lon)
-        _sqs_enqueue_msg(q, m)
+        _msg_build(SQS_BOOT, lat, lon)
     except (Exception, ) as ex:
         lg.a('error sqs_msg_ddh_booted: {}'.format(ex))
 
 
 def sqs_notify_ble_scan_exception(lat, lon):
     try:
-        q = _sqs_get_queue(WS_QUEUE_IN_NAME)
-        global g_last_sqs_notify_error_ble_hw
-        now = time.perf_counter()
-        if g_last_sqs_notify_error_ble_hw + 86400 > now:
-            return
-        g_last_sqs_notify_error_ble_hw += 86400
-        lg.a('<- DDS BLE hardware error')
-        m = _sqs_build_json_msg(WS_DDH_ERROR_BLE_HARDWARE, lat, lon)
-        _sqs_enqueue_msg(q, m)
+        _msg_build(SQS_ERROR_BLE_HW, lat, lon)
     except (Exception, ) as ex:
         lg.a('error sqs_msg_ddh_error_ble_hardware: {}'.format(ex))
 
@@ -113,38 +132,47 @@ def sqs_notify_ble_scan_exception(lat, lon):
 def _sqs_msg_logger_template(op, mac, lat, lon):
     try:
         lg_sn = dds_get_json_mac_dns(mac)
-        q = _sqs_get_queue(WS_QUEUE_IN_NAME)
         s = '{}/{}/{}'.format(op, lg_sn, mac)
-        m = _sqs_build_json_msg(s, lat, lon)
-        print('sqs {} <- {}'.format(op, s))
-        return _sqs_enqueue_msg(q, m)
+        _msg_build(s, lat, lon)
     except (Exception, ) as ex:
         lg.a('error sqs_msg_logger_template: {}'.format(ex))
 
 
-def sqs_notify_logger_error(mac, lat, lon):
-    lg_sn = dds_get_json_mac_dns(mac)
-    op = WS_LOGGER_ERRORS_MAX
-    return _sqs_msg_logger_template(op, mac, lg_sn, lat, lon)
+def sqs_notify_logger_max_errors(mac, lat, lon):
+    op = SQS_LOGGER_MAX_ERRORS
+    return _sqs_msg_logger_template(op, mac, lat, lon)
 
 
 def sqs_notify_oxygen_zeros(mac, lat, lon):
-    lg_sn = dds_get_json_mac_dns(mac)
-    op = WS_LOGGER_ERROR_OXYGEN
-    return _sqs_msg_logger_template(op, mac, lg_sn, lat, lon)
+    op = SQS_LOGGER_ERROR_OXYGEN
+    return _sqs_msg_logger_template(op, mac, lat, lon)
 
 
 def sqs_notify_logger_download(mac, lat, lon):
-    lg_sn = dds_get_json_mac_dns(mac)
-    op = WS_LOGGER_DOWNLOAD
-    return _sqs_msg_logger_template(op, mac, lg_sn, lat, lon)
+    op = SQS_LOGGER_DL
+    return _sqs_msg_logger_template(op, mac, lat, lon)
 
 
-# testing
-def main():
-    print(os.getenv('DDH_AWS_KEY_ID'))
-    sqs_notify_ddh_booted('my_lat', 'my_lon')
+def sqs_serve():
+    q = sqs.get_queue_by_name(QueueName=SQS_UP_QUEUE_NAME)
+    fol = get_dds_folder_path_sqs()
+    files = glob.glob('{}/*.sqs'.format(fol))
+    for _ in files:
+        f = open(_, 'r')
+        d = json.load(f)
+        try:
+            lg.a('serving file {}'.format(_))
+            _msg_queue(q, json.dumps(d))
+            # delete SQS file
+            os.unlink(_)
+        except (Exception, ) as ex:
+            lg.a('error sqs_serve: {}'.format(ex))
+        finally:
+            if f:
+                f.close()
 
 
 if __name__ == '__main__':
-    main()
+    sqs_notify_ddh_booted('my_lat', 'my_lon')
+    sqs_serve()
+    sqs_ws()
